+++ src/characters/NeptuneGrass.ts
@@ -5,6 +5,7 @@ class NeptuneGrass extends Character {
   protected readonly _isPrey: boolean = true;
   protected readonly _isObstacle: boolean = false;
   protected readonly _isBenthic: boolean = true;
+  protected readonly _isSolitary: boolean = false;
   protected readonly _imageFilename: string = "neptune-grass.png";
   protected readonly _foodValue: number = 5;
   protected readonly _damage: number = 0;