+++ src/characters/PlasticBag.ts
@@ -5,6 +5,7 @@ class PlasticBag extends Character {
   protected readonly _isPrey: boolean = false;
   protected readonly _isObstacle: boolean = true;
   protected readonly _isBenthic: boolean = false;
+  protected readonly _isSolitary: boolean = true;
   protected readonly _imageFilename: string = "plastic-bag.png";
   protected readonly _foodValue: number = 0;
   protected readonly _damage: number = 8;