+++ src/characters/Sardine.ts
@@ -5,6 +5,7 @@ class Sardine extends Character {
   protected readonly _isPrey: boolean = true;
   protected readonly _isObstacle: boolean = false;
   protected readonly _isBenthic: boolean = false;
+  protected readonly _isSolitary: boolean = false;
   protected readonly _imageFilename: string = "sardine.png";
   protected readonly _foodValue: number = 0.6;
   protected readonly _damage: number = 0;