+++ src/characters/Shrimp.ts
@@ -5,6 +5,7 @@ class Shrimp extends Character {
   protected readonly _isPrey: boolean = true;
   protected readonly _isObstacle: boolean = false;
   protected readonly _isBenthic: boolean = false;
+  protected readonly _isSolitary: boolean = true;
   protected readonly _imageFilename: string = "shrimp.png";
   protected readonly _foodValue: number = 0.01;
   protected readonly _damage: number = 0;